I have created a README.es.md file for the Spanish translation.
This file contains the Spanish translation of README.md.

Not only did I do the translation, I also added a small icon to be
able to jump from the README.md from Spanish to English.

I also modified the README.md to be able to jump from the English
README.md to the Spanish translation.